Update README.md
In this lab, we use Bayesian regression to t a linear model. Consider a linear model of the form

z = a1x + a0 + w;

(1)

where x is the scale input variable, and a = (a0; a1)T is the vector-valued parameter with unknown entries a0, a1, and w is the additive Gaussian noise:

w N (0; 2);

(2)

where 2 is a known parameter.

Suppose that we have access to a training data set containing N samples fx1; z1 g; fx2; z2g; : : : ; fxN ; zN g. We aim to estimate the parameter a by nding its posterior distribution. When the training nishes, we make predictions based on new inputs. We consider a Bayesian approach, which models the parameter a as a zero mean isotropic Gaussian random vector whose probability distribution is expressed as